+++ src/rpc-handlers/index.js
@@ -1,9 +1,7 @@
 const git = require('./git')
 const acl = require('./acl')
-const bpr = require('./bpr')
 
 module.exports = {
   git,
   acl,
-  bpr,
 }
